pyproject.toml
[tool.poetry]
name = "fastapi_jwt_oauth"
version = "0.0.1"
description = "FastAPI JWT OAuth"
authors = ["williamqosmos <max.2898@yandex.ru>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.11.1"
fastapi = {extras = ["all"], version = "^0.115.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = ">=3.1.0"
sqlalchemy = "^2.0.19"
pydantic = "^2.0.3"
asyncpg = "^0.28.0"
dishka = "^1.1.1"
uvicorn = "^0.29.0"
redis = "^5.2.0"
fastapi-oauth2 = "^1.3.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
mypy = "^1.4.1"
black = "^23.7.0"
pytest = "^7.4.0"
pre-commit = "^3.3.3"
deptry = "^0.20.0"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/venv
/.venv
/__pycache__
/.mypy_cache
/.pytest_cache
/db.sqlite3
/alembic
'''

[tool.ruff]
line-length = 120
select = ["E", "F", "UP", "I"]
ignore = ["E501"]
fixable = ["A", "B", "C", "D", "E", "F", "I"]
exclude = [
    ".direnv",
    ".eggs",
    ".git",
    ".mypy_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "migrations",
    "__pypackages__",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "alembic"
]

[tool.deptry]
ignore = ["DEP003"]

[tool.deptry.per_rule_ignores]
DEP002 = ["asyncpg", "uvicorn"]

[tool.ruff.isort]
section-order = ["future", "fastapi", "standard-library", "third-party",  "first-party", "local-folder"]

[tool.ruff.isort.sections]
fastapi = ["fastapi"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
